requirements.auth.txt
djangosaml2==1.9.*
django-defender==0.9.*